Fixed Collision Layers and Damage Flash. Fixed Scenes with new Player Prefab.
