+++ Assets/Game/HealthSystem/Scripts/DamageFlash.cs
@@ -2,45 +2,49 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class DamageFlash : MonoBehaviour
+namespace DLSU.SpacePirates.HealthSystem
 {
-    [SerializeField]
-    private SpriteRenderer spriteRenderer;
-    [SerializeField]
-    private Color flashColor;
-    [SerializeField]
-    private float flashTime;
-    private float ticks = 0.0f;
-    private bool damaged = false;
-    private Color oldColor;
-    // Start is called before the first frame update
-    void Start()
+    public class DamageFlash : MonoBehaviour
     {
-        oldColor = spriteRenderer.color;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        if (damaged)
+        [SerializeField]
+        private SpriteRenderer spriteRenderer;
+        [SerializeField]
+        private Color flashColor;
+        [SerializeField]
+        private float flashTime;
+        private float ticks = 0.0f;
+        private bool damaged = false;
+        private Color oldColor;
+        // Start is called before the first frame update
+        void Start()
         {
-            ticks += Time.deltaTime;
-            spriteRenderer.color = Color.Lerp(flashColor, oldColor, ticks / flashTime);
-            if (ticks >= flashTime)
+            oldColor = spriteRenderer.material.color;
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+            if (damaged)
             {
-                spriteRenderer.color = oldColor;
-                damaged = false;
+                ticks += Time.deltaTime;
+                spriteRenderer.material.color = Color.Lerp(flashColor, oldColor, ticks / flashTime);
+                if (ticks >= flashTime)
+                {
+                    spriteRenderer.color = oldColor;
+                    damaged = false;
+                }
+            }
+        }
+
+        public void OnHealthChanged(int oldHealth, int newHealth)
+        {
+            if (oldHealth > newHealth)
+            {
+                spriteRenderer.material.color = flashColor;
+                damaged = true;
+                ticks = 0.0f;
             }
         }
     }
 
-    public void OnHealthChanged(int oldHealth, int newHealth)
-    {
-        if (oldHealth > newHealth)
-        {
-            spriteRenderer.color = flashColor;
-            damaged = true;
-            ticks = 0.0f;
-        }
-    }
 }
