+++ Assets/Game/WeaponSystem/Prefabs/PeaShooterProjectile.prefab
@@ -15,7 +15,8 @@ GameObject:
   - component: {fileID: 2653843558156649762}
   - component: {fileID: 876286404105053798}
   - component: {fileID: 3018747958972852830}
-  m_Layer: 11
+  - component: {fileID: 6941198195632472935}
+  m_Layer: 12
   m_Name: PeaShooterProjectile
   m_TagString: PlayerLaser
   m_Icon: {fileID: 0}
@@ -179,3 +180,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e08e675c8c828f546ae9d1e72165e6e4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &6941198195632472935
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6671932444758613728}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8dc3b472907e4be4abf24d63a337e130, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  damage: 5
