--- Assets/Game/WeaponSystem/Scripts/PeaShooterProjectileMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 00c5381404a649e459bea9b669e4e448
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
