--- Assets/Game/Weapons/Sprites.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2264f740f8452364192866ff2f46c7d1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
