+++ Assets/Resources/Prefabs/Asteroid.prefab
@@ -18,6 +18,7 @@ GameObject:
   - component: {fileID: 5776598680739971936}
   - component: {fileID: 1129857079084611536}
   - component: {fileID: 6230135206761991813}
+  - component: {fileID: 8564726654025874329}
   m_Layer: 10
   m_Name: Asteroid
   m_TagString: Obstacle
@@ -299,3 +300,16 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   activeEnemyDeath: {fileID: 11400000, guid: 0d52f4c9d8ba5ab4fb669fac61828cef, type: 2}
+--- !u!114 &8564726654025874329
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6212950485000574065}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8dc3b472907e4be4abf24d63a337e130, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  damage: 1
