+++ ProjectSettings/TagManager.asset
@@ -21,8 +21,8 @@ TagManager:
   - CameraCollision
   - EnemyProjectile
   - Enemy
-  - 
-  - 
+  - Player
+  - PlayerProjectile
   - 
   - 
   -
